requirements.txt
streamlit
openai
langchain
langchain_community
faiss-cpu
langchain_openai
pypdf
python-dotenv
pyngrok
